move codemirror extensions into their own folder
