+++ client/components/codeEditor/extensions/autocompleteEmoji.js
