+++ client/components/codeEditor/extensions/customFolding.js
