+++ client/components/codeEditor/extensions/customHighlight.js
