+++ client/components/codeEditor/extensions/legacyCustomHighlight.js
